Fix rash at 0 timer.

+++ src/server/game/Game.cpp
@@ -114,16 +114,16 @@ void Game::updateShips(std::shared_ptr<Server> server, Parser::entity_t entity)
                 for (int i = 0; i < communication.second.nbrItems; i++) {
                     switch (communication.second.event[i]) {
                         case Communication::EventInput::Key_up:
-                            entity.instance["y"] = std::any_cast<float>(entity.instance["y"]) - 1 * std::any_cast<float>(entity.instance["speed"]);
+                            entity.instance["y"] = std::any_cast<int>(entity.instance["y"]) - 1 * std::any_cast<int>(entity.instance["speed"]);
                             break;
                         case Communication::EventInput::Key_down:
-                            entity.instance["y"] = std::any_cast<float>(entity.instance["y"]) + 1 * std::any_cast<float>(entity.instance["speed"]);
+                            entity.instance["y"] = std::any_cast<int>(entity.instance["y"]) + 1 * std::any_cast<int>(entity.instance["speed"]);
                             break;
                         case Communication::EventInput::Key_left:
-                            entity.instance["x"] = std::any_cast<float>(entity.instance["x"]) - 1 * std::any_cast<float>(entity.instance["speed"]);
+                            entity.instance["x"] = std::any_cast<int>(entity.instance["x"]) - 1 * std::any_cast<int>(entity.instance["speed"]);
                             break;
                         case Communication::EventInput::Key_right:
-                            entity.instance["x"] = std::any_cast<float>(entity.instance["x"]) + 1 * std::any_cast<float>(entity.instance["speed"]);
+                            entity.instance["x"] = std::any_cast<int>(entity.instance["x"]) + 1 * std::any_cast<float>(entity.instance["speed"]);
                             break;
                         default:
                             break;
@@ -142,7 +142,7 @@ void Game::updateColisions(std::shared_ptr<Server> server, Parser::entity_t enti
         for (auto &entity_colision : _entities) {
             if (entity_colision.type == "__tile__") {
                 // HANDLE COLLISION
-                if (std::any_cast<float>(entity.instance["x"]) == std::any_cast<float>(entity_colision.instance["x"]) && std::any_cast<float>(entity.instance["y"]) == std::any_cast<float>(entity_colision.instance["y"])) {
+                if (std::any_cast<int>(entity.instance["x"]) == std::any_cast<int>(entity_colision.instance["x"]) && std::any_cast<int>(entity.instance["y"]) == std::any_cast<int>(entity_colision.instance["y"])) {
                     std::cout << "COLISION" << std::endl;
                     if (std::any_cast<int>(entity.instance["hp"]) != 0) {
                         entity.instance["hp"] = 0;
@@ -151,7 +151,7 @@ void Game::updateColisions(std::shared_ptr<Server> server, Parser::entity_t enti
             }
             else if (entity_colision.type == "missile" && entity.type != "missile") {
                 if (std::any_cast<int>(entity.instance["id"]) != std::any_cast<int>(entity_colision.instance["id"])) {
-                    if (std::any_cast<float>(entity.instance["x"]) == std::any_cast<float>(entity_colision.instance["x"]) && std::any_cast<float>(entity.instance["y"]) == std::any_cast<float>(entity_colision.instance["y"])) {
+                    if (std::any_cast<int>(entity.instance["x"]) == std::any_cast<int>(entity_colision.instance["x"]) && std::any_cast<int>(entity.instance["y"]) == std::any_cast<int>(entity_colision.instance["y"])) {
                         std::cout << "COLISION" << std::endl;
                         if (std::any_cast<int>(entity.instance["hp"]) != 0) {
                             entity.instance["hp"] = 0;
@@ -166,21 +166,21 @@ void Game::updateColisions(std::shared_ptr<Server> server, Parser::entity_t enti
 void Game::updateEntities(std::shared_ptr<Server> server, Parser::entity_t entity)
 {
     if (entity.type == "missile") {
-        entity.instance["x"] = std::any_cast<float>(entity.instance["x"]) + std::any_cast<float>(entity.instance["direction_x"]) * std::any_cast<float>(entity.instance["speed"]);
-        entity.instance["y"] = std::any_cast<float>(entity.instance["y"]) + std::any_cast<float>(entity.instance["direction_y"]) * std::any_cast<float>(entity.instance["speed"]);
+        entity.instance["x"] = std::any_cast<int>(entity.instance["x"]) + std::any_cast<int>(entity.instance["direction_x"]) * std::any_cast<int>(entity.instance["speed"]);
+        entity.instance["y"] = std::any_cast<int>(entity.instance["y"]) + std::any_cast<int>(entity.instance["direction_y"]) * std::any_cast<int>(entity.instance["speed"]);
     }
     if (entity.type != "player" || entity.type != "missile") {
-        entity.instance["x"] = std::any_cast<float>(entity.instance["x"]) - 1;
+        entity.instance["x"] = std::any_cast<int>(entity.instance["x"]) - 1;
     }
     if (entity.type == "player") {
-        _ships.push_back(std::make_shared<Ship>(Communication::Position{std::any_cast<float>(entity.instance["x"]), std::any_cast<float>(entity.instance["y"])}, std::any_cast<int>(entity.instance["id"]), ShipType::PLAYER));
+        _ships.push_back(std::make_shared<Ship>(Communication::Position{std::any_cast<float>(entity.instance["x"]), std::any_cast<float>(entity.instance["y"])}, std::any_cast<float>(entity.instance["id"]), ShipType::PLAYER));
 
         _entities.push_back(static_cast<Parser::entity_t>(
             Parser::entity_t({"missile",
                               {std::vector<std::string>({})},
                               std::map<std::string, std::any>{
-                                  {"x", std::any_cast<float>(entity.instance["x"]) + 1},
-                                  {"y", std::any_cast<float>(entity.instance["y"])},
+                                  {"x", std::any_cast<int>(entity.instance["x"]) + 1},
+                                  {"y", std::any_cast<int>(entity.instance["y"])},
                                   {"speed", 2},
                                   {"id", std::any_cast<int>(entity.instance["id"])},
                                   {"direction_x", 1.0},
